fees and login

+++ src/GetAllStudentGui.java
@@ -8,8 +8,8 @@ import java.util.ArrayList;
 public class GetAllStudentGui extends JFrame implements ActionListener {
     Container container;
     JButton updateButton, deleteButton, showStudentDetailsButton, backButton, chooseDateButton;
-    JTextField nameTextField, phoneNumberTextField, registrationDateTextField;
-    JLabel nameLabel, phoneNumberLabel, registrationDateLabel, titleLabel;
+    JTextField nameTextField, phoneNumberTextField, registrationDateTextField,feesTextField;
+    JLabel nameLabel, phoneNumberLabel, registrationDateLabel, titleLabel, feesLabel;
     MainGui prop;
 
     DefaultTableModel model;
@@ -63,8 +63,8 @@ public class GetAllStudentGui extends JFrame implements ActionListener {
         chooseDateButton.setForeground(Theme.blackColor);
         chooseDateButton.setFont(Theme.fontSmallBold);
 
-        updateButton.setBounds(600, 280, 141, 28);
-        deleteButton.setBounds(800, 280, 141, 28);
+        updateButton.setBounds(600, 290, 141, 28);
+        deleteButton.setBounds(800, 290, 141, 28);
         backButton.setBounds(800, 610, 141, 28);
         showStudentDetailsButton.setBounds(550, 610, 191, 28);
         chooseDateButton.setBounds(894, 183, 50, 28);
@@ -99,15 +99,21 @@ public class GetAllStudentGui extends JFrame implements ActionListener {
         titleLabel.setFont(Theme.bigFont);
         titleLabel.setForeground(Theme.blackColor);
 
+        feesLabel = new JLabel("Fees");
+        feesLabel.setFont(Theme.smallFont);
+        feesLabel.setForeground(Theme.blackColor);
+
         nameLabel.setBounds(78, 128, 176, 28);
         phoneNumberLabel.setBounds(78, 183, 176, 28);
         registrationDateLabel.setBounds(528, 183, 176, 28);
+        feesLabel.setBounds(78, 232, 176, 28);
         titleLabel.setBounds(350, 25, 554, 52);
 
         container.add(nameLabel);
         container.add(phoneNumberLabel);
         container.add(registrationDateLabel);
         container.add(titleLabel);
+        container.add(feesLabel);
 
         //JTextField
         nameTextField = new JTextField();
@@ -125,13 +131,21 @@ public class GetAllStudentGui extends JFrame implements ActionListener {
         registrationDateTextField.setForeground(Theme.blackColor);
         registrationDateTextField.setFont(Theme.inputColor);
 
+        feesTextField = new JTextField();
+        feesTextField.setBackground(Theme.whiteColor);
+        feesTextField.setForeground(Theme.blackColor);
+        feesTextField.setFont(Theme.inputColor);
+
+
         nameTextField.setBounds(225, 128, 400, 28);
         phoneNumberTextField.setBounds(225, 183, 230, 28);
         registrationDateTextField.setBounds(654, 183, 230, 28);
+        feesTextField.setBounds(225, 232, 230, 28);
 
         container.add(nameTextField);
         container.add(phoneNumberTextField);
         container.add(registrationDateTextField);
+        container.add(feesTextField);
 
         //JTable and JScroll
         model = new DefaultTableModel();
@@ -145,6 +159,7 @@ public class GetAllStudentGui extends JFrame implements ActionListener {
         model.addColumn("Full Name");
         model.addColumn("Phone Number");
         model.addColumn("Register Date");
+        model.addColumn("Fees");
         studentsTable.setFont(new Font("Cleaver's_Juvenia_Bloc...",Font.PLAIN , 18));
 
         ArrayList<String> studentsArrayList = FileOperation.storage("students.txt");
@@ -154,17 +169,20 @@ public class GetAllStudentGui extends JFrame implements ActionListener {
             String[] rowSubSub1 = rowSub[1].split(":");
             String[] rowSubSub2 = rowSub[2].split(":");
             String[] rowSubSub3 = rowSub[3].split(":");
+            String[] rowSubSub4 = rowSub[4].split(":");
 
-            model.addRow(new Object[]{rowSubSub1[1], rowSubSub2[1], rowSubSub3[1]});
+            model.addRow(new Object[]{rowSubSub1[1], rowSubSub2[1], rowSubSub3[1],rowSubSub4[1]});
         }
         DefaultTableCellRenderer cellRenderer;
         studentsTable.getColumnModel().getColumn(0).setPreferredWidth(400);
         studentsTable.getColumnModel().getColumn(1).setPreferredWidth(300);
         studentsTable.getColumnModel().getColumn(2).setPreferredWidth(300);
+        studentsTable.getColumnModel().getColumn(3).setPreferredWidth(300);
         cellRenderer = new DefaultTableCellRenderer();
         cellRenderer.setHorizontalAlignment(JLabel.CENTER);
         studentsTable.getColumnModel().getColumn(1).setCellRenderer(cellRenderer);
         studentsTable.getColumnModel().getColumn(2).setCellRenderer(cellRenderer);
+        studentsTable.getColumnModel().getColumn(3).setCellRenderer(cellRenderer);
 
         studentsTable.addMouseListener(new MouseAdapter() {
             @Override
@@ -206,10 +224,11 @@ public class GetAllStudentGui extends JFrame implements ActionListener {
                 try {
                     String[] subSubIDRow = subRow[0].split(":");
                     String[] subSubNameRow = subRow[1].split(":");
+                    String[] subSubFeesRow = subRow[4].split(":");
 
 
                     //get Details for user
-                    getDetailsUser = new GetDetailsUser(this, subSubIDRow[1], subSubNameRow[1]);
+                    getDetailsUser = new GetDetailsUser(this, subSubIDRow[1], subSubNameRow[1],subSubFeesRow[1]);
                     getDetailsUser.setVisible(true);
 
                 } catch (Exception exception) {
@@ -232,22 +251,25 @@ public class GetAllStudentGui extends JFrame implements ActionListener {
                 String name = nameTextField.getText().trim();
                 String phone = phoneNumberTextField.getText().trim();
                 String date = registrationDateTextField.getText().trim();
+                String fees = feesTextField.getText().trim();
 
-                if (Operation.isPhoneNumber(phone) && Operation.isDate(date) && !name.isEmpty() && !phone.isEmpty() && !date.isEmpty() && Operation.checkIsNotHash(name) && Operation.checkIsNotHash(date) && Operation.checkIsNotHash(phone)) {
+                if (Operation.isPhoneNumber(phone) && Operation.isDate(date) && !name.isEmpty() && !phone.isEmpty() && !date.isEmpty() && !fees.isEmpty() && Operation.checkIsNotHash(name) && Operation.checkIsNotHash(date) && Operation.checkIsNotHash(phone) &&Operation.checkIsNotHash(fees) && Operation.isNumeric(fees)) {
 
 
                     studentDefaltTableModel.setValueAt(name, studentsTable.getSelectedRow(), 0);
                     studentDefaltTableModel.setValueAt(phone, studentsTable.getSelectedRow(), 1);
                     studentDefaltTableModel.setValueAt(date, studentsTable.getSelectedRow(), 2);
+                    studentDefaltTableModel.setValueAt(fees, studentsTable.getSelectedRow(), 3);
 
                     JOptionPane.showMessageDialog(null, "Student Update Successfully");
                     nameTextField.setText("");
                     phoneNumberTextField.setText("");
                     registrationDateTextField.setText("");
+                    feesTextField.setText("");
 
                     ArrayList<String> studentsArrayList = FileOperation.storage("students.txt");
                     int index = 0;
-                    String subRow[] = new String[4];
+                    String subRow[] = new String[5];
                     for (String row : studentsArrayList) {
                         if (row.contains(oldName)) {
                             subRow = row.split("#");
@@ -256,7 +278,7 @@ public class GetAllStudentGui extends JFrame implements ActionListener {
                         }
                     }
                     String[] subSubIDRow = subRow[0].split(":");
-                    studentsArrayList.set(index, "id:" + subSubIDRow[1] + "#" + "name:" + name + "#" + "phone:" + phone + "#" + "date:" + date);
+                    studentsArrayList.set(index, "id:" + subSubIDRow[1] + "#" + "name:" + name + "#" + "phone:" + phone + "#" + "date:" + date + "#" + "fees:" + fees);
                     FileOperation.writeToFile("students.txt", studentsArrayList);
                 } else {
                     JOptionPane.showMessageDialog(null, "FullName Field Must Be a text, Phone Field Must Be a Number, Date Field Must Be a Date", "Invalid Entry", 0);
@@ -275,22 +297,26 @@ public class GetAllStudentGui extends JFrame implements ActionListener {
             DefaultTableModel studentDefaltTableModel = (DefaultTableModel) studentsTable.getModel();
 
             if (studentsTable.getSelectedRowCount() == 1) {
-                studentDefaltTableModel.removeRow(studentsTable.getSelectedRow());
-                FileOperation.deletePaymentAndLevelForStudent(oldName);
-                ArrayList<String> studentsArrayList = FileOperation.storage("students.txt");
-                int index = 0;
-                for (String row : studentsArrayList) {
-                    if (row.contains(oldName)) {
-                        index = studentsArrayList.indexOf(row);
-                        break;
+                int option = JOptionPane.showConfirmDialog(null,"Are You Sure To Delete This Student?","Delete",JOptionPane.YES_NO_OPTION);
+                if(option == 0) {
+                    studentDefaltTableModel.removeRow(studentsTable.getSelectedRow());
+                    FileOperation.deletePaymentAndLevelForStudent(oldName);
+                    ArrayList<String> studentsArrayList = FileOperation.storage("students.txt");
+                    int index = 0;
+                    for (String row : studentsArrayList) {
+                        if (row.contains(oldName)) {
+                            index = studentsArrayList.indexOf(row);
+                            break;
+                        }
                     }
+                    studentsArrayList.remove(index);
+                    FileOperation.writeToFile("students.txt", studentsArrayList);
+                    JOptionPane.showMessageDialog(null, "Student Delete Successfully");
+                    nameTextField.setText("");
+                    phoneNumberTextField.setText("");
+                    registrationDateTextField.setText("");
+                    feesTextField.setText("");
                 }
-                studentsArrayList.remove(index);
-                FileOperation.writeToFile("students.txt", studentsArrayList);
-                JOptionPane.showMessageDialog(null, "Student Delete Successfully");
-                nameTextField.setText("");
-                phoneNumberTextField.setText("");
-                registrationDateTextField.setText("");
             } else {
                 if (studentsTable.getRowCount() == 0) {
                     JOptionPane.showMessageDialog(null, "The Table Is Empty");
@@ -310,11 +336,13 @@ public class GetAllStudentGui extends JFrame implements ActionListener {
         oldName = name;
         String phone = studentDefaltTableModel.getValueAt(studentsTable.getSelectedRow(), 1).toString();
         String date = studentDefaltTableModel.getValueAt(studentsTable.getSelectedRow(), 2).toString();
+        String fees = studentDefaltTableModel.getValueAt(studentsTable.getSelectedRow(), 3).toString();
 
         //set to textField
         nameTextField.setText(name);
         phoneNumberTextField.setText(phone);
         registrationDateTextField.setText(date);
+        feesTextField.setText(fees);
 
     }
 }
